% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_best_item_subsets}
\alias{get_best_item_subsets}
\title{Get Best Item Subsets
This is the main function of the program}
\usage{
get_best_item_subsets(
  items,
  data,
  actual_scores,
  scale = 5,
  resamples = 0,
  quantile_groups = 4
)
}
\arguments{
\item{items}{vector of column names (strings)}

\item{data}{data.frame from which to pull values}

\item{actual_scores}{actual scores to compare performance to}

\item{scale}{The size of the scale to be used. Defualts to 5.}

\item{resamples}{The number of resample replicates by which to compute errors}

\item{quantile_groups}{The number of quantile groups across which to compute the average error. Defaults to 4.}
}
\description{
Get Best Item Subsets
This is the main function of the program
}
