% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_best_n_items}
\alias{get_best_n_items}
\title{Get best n items
Uses the maximum average error across quantiles as the performance metric. Employs exhaustive search across item combinations. Is predictably slow for large numbers of items.}
\usage{
get_best_n_items(
  n,
  items,
  data,
  actual_scores,
  scale = 5,
  quantile_groups = 4,
  resamples = 0
)
}
\arguments{
\item{n}{The size of the subset to choose.}

\item{items}{vector of column names (strings)}

\item{data}{data.frame from which to pull values}

\item{actual_scores}{actual scores to compare performance to}

\item{scale}{The size of the scale to be used. Defualts to 5.}

\item{quantile_groups}{The number of quantile groups across which to compute the average error. Defaults to 4.}

\item{resamples}{The number of resample replicates by which to compute errors. Defaults to 0.}
}
\description{
Get best n items
Uses the maximum average error across quantiles as the performance metric. Employs exhaustive search across item combinations. Is predictably slow for large numbers of items.
}
