% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{reorganize_output}
\alias{reorganize_output}
\title{Reorganize Output
This function is for internal use within get_best_item_subsets. Perhaps it should not be a function.}
\usage{
reorganize_output(best_item_subsets_result)
}
\arguments{
\item{best_item_subsets_result}{Again, this is for internal use.
reorganize_output()}
}
\description{
Reorganize Output
This function is for internal use within get_best_item_subsets. Perhaps it should not be a function.
}
